FLIZ-388 refactor: firebase 및 firebase/messaging 로직 리팩토링
